requirements.txt
opencv-python
numpy
python-dotenv
fastapi
uvicorn[standard]
